Split out more android specific classes
